Adds a basic offline check to InstallableManager.
Adds a check for existence of a fetch handler in the service worker as a small
step towards checking offline capabilities for the A2HS feature.

Modifies CheckHasServiceWorker method to return an enum distinguishing a
service worker with and without a registered fetch handler.

BUG=601337

Review-Url: https://codereview.chromium.org/2751343002
Cr-Original-Commit-Position: refs/heads/master@{#458643}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3e32ed5683dd6e1cf0178e07c6e58b0e8198a7c6