Update AutoImport to import nothing by default (in absence of policy and master_prefs).
As decided in conclusion of experiment https://crbug.com/555550#c114

This simplifies the code a lot, dropping all the weird defaults accrued
over the years. AutoImport is now strictly driven by policies + master_prefs.

Also deprecated the distribution entries for the import prefs which were
duplicates of the prefs:: entries... added legacy mapping in one place in code
and it then simplifies the rest of the logic (the kDistroDict is never
registered so this was somewhat of a pain).

The new code also makes it trivial to fix issue 601697 :) -- done!
(i.e. added support for auto import of autofill form data and saved passwords
 although actual support will depend on the support of the importer
 for the imported browser -- which is the same as the support from a
 manual import).

BUG=555550, 601697

Review-Url: https://codereview.chromium.org/2705113005
Cr-Original-Commit-Position: refs/heads/master@{#458853}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dac09c03b9f3cc44adf71e8836f6bebabec7a61d