Implement the BackgroundFetch{Fail,ed} Service Worker events
This CL implements the two final events part of the Background Fetch
API: backgroundfetchfail and backgroundfetched. They'll be dispatched
from the BackgroundFetchEventDispatcher, through the
ServiceWorkerEventDispatcher, to the renderer, where they'll run.

The BackgroundFetchSettledFetch structure has been defined in the
SWEventDispatcher's mojom file, but will move to its more appropriate
location in Blink when the ServiceWorkerFetchRequest/Response types
have been migrated to Mojo proper. This is in-progress.

BUG=692534, 692581

Review-Url: https://codereview.chromium.org/2767093004
Cr-Original-Commit-Position: refs/heads/master@{#459469}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4b613c4cd7c83a70d4f176bcb6394b2a800bbd4d