Add a UMA counter for tap disambiguation result.
This is to collect data informing a possible removal of this feature.
In particular, only if the user tapped inside the popup can it be
considered to be a (possibly) successful UX; if the user dismissed it by
tapping outside or pressing the back button, then it can be considered
to have caused more harm than good.

BUG=705117

Review-Url: https://codereview.chromium.org/2775013002
Cr-Original-Commit-Position: refs/heads/master@{#460013}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 642318401fc47bea352b860ab8786854abb7c314