[IndexedDB] Abort transaction on unknown blobs instead of crashing
There is a current refcounting known issue where a blob can be sent to
another process but while that IPC is scheduled the source process
dies. The blob will be non-existent. We want to gracefully handle this
error in IndexedDB instead of crashing.

(also fixes missing include in the connection class)

BUG=677780

Review-Url: https://codereview.chromium.org/2779433002
Cr-Original-Commit-Position: refs/heads/master@{#460491}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2a0b10eb79e78d3aee52538d66768381769469dc