Revert "Revert "[Home] Add some histograms for Chrome Home""
This reverts commit 826f594306b79439714faec487b06a9fd860c4a2.

The fix is to check if the tab containing the NTP is already showing
(e.g. from a back navigation or because it's the selected tab on
startup) and call onNewTabPageShown() if it is rather than relying
on a call to TabObserver#onShown().

BUG=700592, 700593
TBR=isherman@chromium.org, mdjones@chromium.org

Review-Url: https://codereview.chromium.org/2788163002
Cr-Original-Commit-Position: refs/heads/master@{#461208}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7ca06a2f0b0cd9ade0dd6a47f4fa3c6605479af0