Refine tap disambiguation UMA to track same-node/different-node.
The Android tap disambiguation feature makes users retap on a zoomed screenshot
when the user tapped on tiny close-together links.  This feature is
currently being considered for removal.  If the user's tap would've been
anyway, in the first place, accurately resolved to the node they
ultimately selected after disambiguation, then the popup UI achieved
nothing and we should consider this a failure case.

Notes:
- Because this must be measured in the renderer and the synthetic
taps sent before had nothing to identify them as relating to
disambiguation, I introduced a new ViewMsg_ResolveTapDisambiguation.
I think this is anyway a better architecture -- this is a highly unusual
scenario not much resembling an ordinary touch gesture, and risks
interacting weirdly with ordinary event routing.
- Although handling this in WebViewImpl is broken with OOPIF, this feature
is already broken with OOPIF.  That's one of the reasons it's considered
for removal.
- I'm using hit node top-left offset relative the document as an
identifier the DOM node itself.  This seems less sketchy than persisting
a pointer, and it's good enough for this use case because the page is
generally static during a tap disambiguation (if it self-scrolls, for
instance, the popup will be dismissed with category OTHER).

BUG=705117

Review-Url: https://codereview.chromium.org/2783233004
Cr-Original-Commit-Position: refs/heads/master@{#461290}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1b55156e5277d39580f4063af924ec7347375ebc