Hide the download summary on file open
Notifications can be dismissed with a
DownloadManager.ACTION_NOTIFICATION_CLICKED intent, which bypasses the
DownloadNotificationService so we don't get a chance to hide the
summary.  This adds a check to the
DownloadBroadcastReceiver#openDownload() path that allows us to hide any
dangling summary notifications that shouldn't be visible.  This also
lets us catch any clicks on the summary which gives us a chance to close
them out as well.

BUG=699323

Review-Url: https://codereview.chromium.org/2796573002
Cr-Original-Commit-Position: refs/heads/master@{#461744}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 803fdf80f842803c7c91dca2a87419afe1b3d60c