Log metrics for the Location Settings Dialog prompt for geolocation.
This prompt is shown when the user has turned off location and a web
site asks for geolocation access. There are various controls to prevent
spamminess (a gesture requirement for non-default search engine sites,
and a backoff system).

This change adds metrics for number of times the dialog was shown,
suppressed (for backoff), accepted or denied. Each metric has two
versions, one for the default search engine and one for other sites.
The value logged for each metric is the current back off level.

This change also cleans up some code for the backoff, using an enum
instead of an int to represent it. It also caches some calculations
done on whether an origin is the default search engine to prevent
it happening quite as much as it was.

BUG=703493

Review-Url: https://codereview.chromium.org/2791193002
Cr-Original-Commit-Position: refs/heads/master@{#462446}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dc48af3732f96f60c30396ed10b158314bc0b8c9