Add metrics for DelayNavigationThrottle
We add 3 'internal' metrics to make sure DelayNavigationThrottle is
adding the expected delays, and to understand the difference between
the specified delay and the actual delay. We expect there to be
differences in cases where the main UI thread is congested, for example.

BUG=704117

Review-Url: https://codereview.chromium.org/2803673002
Cr-Original-Commit-Position: refs/heads/master@{#462626}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5339a7833dc51eae11a0e5b2f29c719122fa79a7