Ensure fixed amount of scrolling in infinite scroll stories.
This patch changes the scrolling logic in the stories such that each
story does the same fixed amount of scrolling in any run.

Apart from that the stories are refactored and unified following the
convention of the system health stories.

PERF SHERIFF NOTE: expect regressions and/or improvements in
- v8.infinite_scroll_tbmv2 (and other variants of this benchmark).
- v8.mobile_infinite_scroll_tbmv2 (and other variants of this benchmark).

BUG=chromium:704124

Review-Url: https://codereview.chromium.org/2773483002
Cr-Original-Commit-Position: refs/heads/master@{#462813}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 95e17ac4602dc38aaa6a44d4f6d2566920d6c75d