Log metrics for the new Default Search Engine geolocation setting.
Recently a new system was introduced which added a preference for
whether the default search engine has geolocation access or not. This
change adds metrics for the state of this setting before the disclosure
UI is shown and after the disclosure UI is shown for the last time.

BUG=700777

Review-Url: https://codereview.chromium.org/2793253002
Cr-Original-Commit-Position: refs/heads/master@{#463204}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5754a56b61f227091f272bc84bfb212b9ba78416