Get signals working in the EXTRA_DATA section of MHTML
The intent of this change is to add some signal data to the bottom of
a MHTML file when generating an offline file if the new OfflinePagesLoadSignalCollecting flag is set.  The signal data
will be analyzed to find a better time to take snapshots, and will eventually be used to decide at runtime when to take snapshots.  Design docs here:
https://docs.google.com/document/d/1AcoZOIb81V68_KDSOCatFSYYMLOTSYj6resg-VAxLgo

BUG=706628

Review-Url: https://codereview.chromium.org/2683493002
Cr-Original-Commit-Position: refs/heads/master@{#463407}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2abcf0413069159c73bceba72f0b6f124021bca5