Add demo-mode for In-Product Help.
This adds a new flag to chrome://flags for controlling a demo mode for
in-product help. By enabling the flag, each in-product help feature
will trigger a single time per session. This is to enable manual testing
of all in-product help UI.

Since enabling this feature can not enable the other in-product help
features, the OnceConditionValidator is changed to not require a feature
to be enabled.

BUG=706309

Change-Id: Id01bc34fb1d297645c115e2f2b13114eb05b2e24
Reviewed-on: https://chromium-review.googlesource.com/473706
Commit-Queue: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#463770}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4fad7069e05d919a504e8db88f6d0a1019aa129b