Allow for pausing timers in background tabs on desktop via about:flags.
This is a first step towards trying to get a flag that's good
enough for us to dogfood ourselves to see if there's something
we could ship regarding pausing background tabs that's web
compatible enough.

Speculatively plumb through the site engagement levels so that the followup
patch can hook them up to the scheduler.

Just shipping this flag as is would clearly not be compatible
enough though.

BUG=709082

Review-Url: https://codereview.chromium.org/2805683003
Cr-Original-Commit-Position: refs/heads/master@{#463749}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 38c67835dc36e9378584fde1da493eccb906db26