Add new builders to the chromium.memory config files.
We're shutting down the chromium.memory.full waterfall and moving
the builders to the chromium.memory waterfall. This CL adds the
entries for the new builders on chromium.memory and should be
landed first.  Once the build-side CL to actually switch the masters
lands, and the masters are restarted, we can land the follow-up
CL to remove the old entries.

R=phajdan.jr@chromium.org, qyearsley@chromium.org
BUG=598859

Review-Url: https://codereview.chromium.org/2817623002
Cr-Original-Commit-Position: refs/heads/master@{#464133}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9acab5dd4495ba6eae78c05659ea09bb601405df