Remove MediaThrottler and MediaThrottleInfoBar
Since the removal of WebMediaPlayerAndroid, we use the
MediaServiceThrottler instead of the MediaThrottler.

This CL cleans up the unused code. It also prevents the
instantiation of BrowserMediaPlayerManager.

A follow up CL to merge BrowserMediaPlayerManager and
RemoteMediaPlayerManager will follow in Q2 2017.

This change frees up ~4.0Kib

BUG=707265, 570711
TEST= built/deployed/ran chrome, made sure nothing crashed.

Review-Url: https://codereview.chromium.org/2794283003
Cr-Original-Commit-Position: refs/heads/master@{#464195}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9542e14c809a1e79e307502b4ed3f00771578b61