Do not play the whole media file before seeking.
I also fixed some comments and fixed pylint line length errors.

This was a bug introduced in crrev/346923003.
Originally we would start playing the video and wait until it got to 1
second, and then we would seek around in it. After crrev/346923003 we
would instead play the entire video and then seek around in it.

sandersd@'s crrev/1962563005 noticed this bug and commented on it, but
we haven't gotten around to fixing it until now.

This change should reduce load on Speed infra machines by reducing the benchmark duration for tough_video_cases_extra from ~5 minutes to ~3 minutes. we get to skip the initial playback for 14 of the pages in tough_video_cases_extra. the video length of test content is around 10 seconds on average, and this change will make us play only 1 second of the video instead of the whole video before we start seeking. So 14*(10-1) = 126 seconds.

Also, I noticed that two of the pages in tough_video_cases are also seeking pages, not time_to_play pages, so this will reduce the runtime for those as well.

In addition to reducing the load, this change will also make seek times become the measurement that they were originally supposed to be. Seek time cold currently doesn't actually measure cold seek time because that part of the video that was seeked to was already played once.

BUG=711125

Review-Url: https://codereview.chromium.org/2813383002
Cr-Original-Commit-Position: refs/heads/master@{#464288}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6a48a45ac9c3138e306f19df0b3fc1974ffe33ec