[Remote suggestions] Report time distribution of triggers
When changing scheduling intervals for soft background fetches, we had
to guess blindly what the impact on traffic this will have.

This CL records the distribution of:
 - when first trigger comes (not necessarily successful), and
 - when the (first) successful trigger comes.

The former helps us to evaluate impact of scheduling changes. The latter
helps us to understand what are the real fetching intervals in the wild.

BUG=710828

Review-Url: https://codereview.chromium.org/2810053004
Cr-Original-Commit-Position: refs/heads/master@{#464595}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 51a10acb782ef180630115d21190b7c70cb44944