Provide instructions if -o or -r is used on public bisect-builds.py
If the user attempts to use the -o or -r options in the public bisect
script, they will fail as unknown options. Instead, we should help
the user get the correct version of the script by pointing to the
page explaining how to configure their environment.

Review-Url: https://codereview.chromium.org/2742003002
Cr-Original-Commit-Position: refs/heads/master@{#464743}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 446bcc3d1881a89471235aff30d25c38b2b3547f