[Push API] Fix crash on null push service at renderer shutdown
https://codereview.chromium.org/2697793004
(6576ecf83ab56518daa1274dfa354aee84c8696b) crashes with a NPE in cases
where there is a stored push subscription but the push service cannot be
accessed.

This wasn't thought to be possible, but it seems it is happening during
renderer shutdown.

This patch cleanly handles that case, and logs it to UMA.

BUG=709950

Review-Url: https://codereview.chromium.org/2814463002
Cr-Original-Commit-Position: refs/heads/master@{#465226}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 82b3bb285fb6944772d7e09afea99a06c10aef67