Don't crash the browser when there's no live Service Worker registration
There are (very few) crash reports where we're trying to get a live Service
Worker registration but fail because there is none. Add a check to make
sure we don't crash the browser, and log UMA so that we can measure how
wide-spread this case is without crashing users.

BUG=704502

Review-Url: https://codereview.chromium.org/2827983002
Cr-Original-Commit-Position: refs/heads/master@{#466104}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a4c052f36399684faf0afe7a8cd243aebfa2cf26