Revamp the Incognito NTP on Desktop
Changes in this CL:

1. Added an about:flag for the "MD Incognito NTP". (Note that the primary
purpose of this change is to improve the text on the Incognito NTP, not
just to materialize it; but the latter results in a shorter flag name and
thematically fits with the other recent MD improvements).

2. Since almost none of the old CSS is valid for the new MD, and since
there are also related HTML changes, we use separate HTML and CSS files.
NTPResourceCache decides which one to serve according to the flag. Once
this fully launches, we will simply remove the previous implementation.

3. The new design follows the CSS rules described in
https://gallery.googleplex.com/projects/MCHbtQVoQ2HCZeKud1YOmPiI.
The requirement that the two sets of bulletpoints determine the width of
the content box has been achieved by JS (the CSS is complicated enough as
it is :-( ).

Tested with the bookmark bar on/off, with a custom theme on/off, and in
front of a UX designer :)

BUG=693525
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:closure_compilation

Review-Url: https://codereview.chromium.org/2804823002
Cr-Original-Commit-Position: refs/heads/master@{#466817}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a8443118aa963ba364ee9687b82d1d9c7eb26af5