Fix disable/enable logic for mobile benchmarks
Some benchmarks use ShouldDisable to not run on android (or run on android).
This CL changes those to use decorators. This is so that we can statically
analyze our benchmarks (in later CLs).

Also adds a unittest for benchmarks with 'mobile' in the name.

Bug: 715177
Change-Id: I611aa0e1924b640ec4a26bf1a8ff31813595daaf
Reviewed-on: https://chromium-review.googlesource.com/477473
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Ulan Degenbaev <ulan@chromium.org>
Reviewed-by: Ned Nguyen <nednguyen@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#467072}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3854d7c235ea528280031646652e030092c0d9b5