Abandon user sign in when policy is retrieved before session started.
If it happens that user policy is retrieved before session is started,
then there's no policy to apply to the user which could be used as
policy enforcement escape. This is a case when some flow went wrong,
and we abandon the user sign in.

BUG=689206

Review-Url: https://codereview.chromium.org/2801993002
Cr-Original-Commit-Position: refs/heads/master@{#468000}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f305a94dd4908c4975db3444e6b80732f48dcb39