[Home] Add expand button and flag to enable it
This change adds an alternate UI for the Chrome Home bottom sheet.
Instead of being able to swipe up when the sheet is peeking, there is
the option to have a button that sits to the left of the omnibox that
expands the sheet when pressed. The feature name for this UI is
"ChromeHomeExpandButton".

This flag runs into the same problems as Chrome Home when first
being enabled -- native is not ready when UI is initialized. To
avoid yet another Chrome restart, the pull-handle version of the UI
will appear briefly before switching to the button.

BUG=716220

Review-Url: https://codereview.chromium.org/2853583002
Cr-Original-Commit-Position: refs/heads/master@{#468781}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d484eacb1a1016a1f2fda3bc57b8e309fd08df76