[subresource_filter] SB throttle can send multiple speculative requests.
This CL adds functionality for sending multiple requests to the
safe browsing database. We utilize this functionality to speculatively
send requests on WillStartRequest and WillRedirectRequest.

In order to evaluate whether speculative checks on every redirect
are necessary, we log metrics for what the navigation delay would look
like if that check was removed.

BUG=671962

Review-Url: https://codereview.chromium.org/2834543003
Cr-Original-Commit-Position: refs/heads/master@{#469073}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 00eefae39baa22b1ad6a1ae37457a1f026d9f40d