Add flag for disabling minimum of ntp tile suggestions
This flag will cause the the client to use history-based suggestions for
NTP tiles even if there are less than 8 suggestions available.

This is particularily useful for Chrome Home and the condensed NTP where
only 4 tiles are used and which does not contain any suggestions until
there would be enough data for 8 suggestions.

BUG=709959

Review-Url: https://codereview.chromium.org/2858883002
Cr-Original-Commit-Position: refs/heads/master@{#469620}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 122c6655af011e18e58f4384e327174dd6212874