Improve linker_verbose_tracking.py
linker_verbose_tracking.py is used to analyze the verbose output from
the VC++ linker to help understand binary-size regressions. However some
of its messages are clear only to the original author, and it completely
fails on the utf-16 files created when redirecting the linker output
from powershell.

This change updates the script to handle both utf-16 and utf-8 files,
and improves the messages printed in a couple of cases.

R=dvallet@chromium.org
BUG=714841

Review-Url: https://codereview.chromium.org/2863843003
Cr-Original-Commit-Position: refs/heads/master@{#469783}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3647b59ef9a6cf01783c7cf7b5ea83eab3296007