[subresource_filter] Implement the "Smart" UI on Android
This patch implements a V1 for the "smart" UI, which just logs a
timestamp every time the UI is shown, and will not show the UI on
subsequent navigations if it is within 30 minutes of a previous impression.

This is implemented using a new WebsiteSetting, which holds a timestamp.

BUG=689992

Review-Url: https://codereview.chromium.org/2795053002
Cr-Original-Commit-Position: refs/heads/master@{#470217}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3f8cdfd3ab1a3e87f90c37f3d7763483ac9e03c4