[Android WebAPKs] Add UMA metric for ShellAPK version
This CL adds a UMA metric for the ShellAPK version. The metric will tell us the
percentage of users with WebAPKs with a given ShellAPK version.

There is a separate metric for "Google Play WebAPKs" and "Browser WebAPKs". We
expect the ShellAPK for "Unbound WebAPKs" to be permanently out of date. We
hope that the ShellAPK for "Browser WebAPKs" stays up to date

BUG=691743

Review-Url: https://codereview.chromium.org/2862493002
Cr-Original-Commit-Position: refs/heads/master@{#470756}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7ff54f9912ee3dd2132e0fc7c1e9bdadc06b46b1