Add a PersistentStore to FeatureEngagementTracker
Currently we use an InMemoryStore for the FeatureEngagementTracker,
which doesn't persist feature Events beyond the scope of the application
lifetime.  However once we move beyond demo mode, the feature engagement
conditions will span months, so we need to track the data for longer
than a single session.

This CL implements the Store interface and backs it with a
leveldb_proto::ProtoDatabase.

BUG=706309

Review-Url: https://codereview.chromium.org/2876633002
Cr-Original-Commit-Position: refs/heads/master@{#471149}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 23b616383acde9c16e0892084498620ecc46856f