Use an MutationObserver to check when a password form disappears after XHR
Right now, we consider a password submitted if the filled form is gone
after an successful XHR or fetch operation. However, if the website
didn't synchronously remove the password form in the event handler, we
will not consider the password submitted.

In that case, we now use a MutationObserver to be notified about future
changes to the form, and if it disappears, we'd still consider the
password submission successful.

BUG=589533
R=dvadym@chromium.org, dglazkov@chromium.org

Review-Url: https://codereview.chromium.org/2865233003
Cr-Original-Commit-Position: refs/heads/master@{#471421}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fcf91a8eaacdeaedb0de1eba39ff4c24fab34323