Report page load timing information for child frames.
This change clarifies that our PLMO::On*Paint callbacks are for the whole page,
and adds support for aggregating timing data across frames in PageLoadTracker.

The current behavior tracks paints across all frames, however this is due to a bug
(crbug.com/705315) and will be fixed soon.

PLMO implementers actually want to track paints across all frames in the page, so
this is desired behavior.

This change adds support for tracking paints across all frames so the current
behavior doesn't regress when 705315 is fixed.

BUG=661702

Review-Url: https://codereview.chromium.org/2859393002
Cr-Original-Commit-Position: refs/heads/master@{#471725}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9f5f6b9548bc4cad0c64d3aa2b5b58ae8dad595b