Sets up form_interactions UKM logger when form is parsed.
We had used the "parse" time and "load" time interchangeably because the
compute time difference between these 2 events were insignificant for
our log events. Because we always parse a form, it is better to
initialize the logger at parse time.

It is safe to change the metric name because the logs have not yet been
consumed in stable release.

BUG=

Review-Url: https://codereview.chromium.org/2883613002
Cr-Original-Commit-Position: refs/heads/master@{#471921}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6031ec2baa3cb70e8105ff2d513a2635839b37ed