Enable store tombstones for swarming bots.
Enable tombstones for swarming bots. Previously if a test case is
crashed on swarming bots, no tombstones will be resolved, and even if
they are attempted to be resolved, they will fail because of the lack of
tool chains.

In this cl, we are enabling tombstones for swarming bots and we are also
adding tool chains to the dependency.

BUG=631213

Review-Url: https://codereview.chromium.org/2883633003
Cr-Original-Commit-Position: refs/heads/master@{#472615}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fc66094f89df0ecca2a3378296304a9c7e66bbff