Metrics for submission event when a user accepts save/update prompt.
On CL https://codereview.chromium.org/2859243004/ it was implemented a UMA metric for heuristics that was used for detecting a form submission in case when this submission was considered to be successful. This CL adds metric subbmission heuristics when the user accepts save/update prompt. So we can understand which submission heuristic is best, and probably which of them are useless.

BUG=None

Review-Url: https://codereview.chromium.org/2890423002
Cr-Original-Commit-Position: refs/heads/master@{#473543}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f8c5824267cdbcb81a45de00e4a8a741cc5f9247